requirements.txt
fastapi
uvicorn[standard]
cohere
requests
pydantic-settings
python-dotenv
streamlit
PyYAML